Moves utility functions to @yaatt/utils
